requirements.txt
numpy==1.19.2
torch==1.7.1
matplotlib==3.3.2
torchvision==0.8.2
Pillow==8.1.0
streamlit==0.76.0
